trabajo terminado

+++ src/funciones.js
@@ -9,7 +9,7 @@ let contador = document.querySelector(".contador");
 
 ////////////////////////////// EXPORTACIONES //////////////////////////////////
 
-import { posTask, getTasks, deleteTask } from "./api.js";
+import { posTask, getTasks, deleteTask, updateTask } from "./api.js";
 
 ////////////////////////////// EXPORTACIONES //////////////////////////////////
 
@@ -19,7 +19,9 @@ import { posTask, getTasks, deleteTask } from "./api.js";
 function validar() {
   let datos = input.value;
   if (datos.length == 0) {
-    alert("campo invalido");
+    Swal.fire({
+      title: "campo vacio",
+    });
   }
 }
 
@@ -57,27 +59,61 @@ function esMayuscula() {
 ////////////////////////////////// FUNCION BTN ELIMINAR /////////////////////////////
 
 function BtnEliminar() {
-  let Eliminar = document.createElement("i");
+  var Eliminar = document.createElement("i");
+  var checkBox = document.querySelector("input");
 
   Eliminar.className = "fa fa-trash";
 
   Eliminar.addEventListener("click", (e) => {
     let item = e.target.parentElement;
-    let checkBox = item.querySelector("input");
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: "btn btn-success",
+        cancelButton: "btn btn-danger",
+      },
+      buttonsStyling: false,
+    });
+    swalWithBootstrapButtons
+      .fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonText: "Yes, delete it!",
+        cancelButtonText: "No, cancel!",
+        reverseButtons: true,
+      })
+      .then((result) => {
+        if (result.isConfirmed) {
+          swalWithBootstrapButtons.fire(
+            "Deleted!",
+            "Your file has been deleted.",
+            "success"
+          );
 
-    if (checkBox.checked) {
-      let cuenta = Number(contador.textContent);
-      cuenta = cuenta - 1;
-      contador.textContent = cuenta;
-    }
-    ul.removeChild(item);
+          if (checkBox.checked) {
+            let cuenta = Number(contador.textContent);
+            cuenta = cuenta - 1;
+            contador.textContent = cuenta;
+          }
+          ul.removeChild(item);
 
-    deleteTask(item.id);
-
-    let items = ul.getElementsByTagName("li");
-    if (items.length == 0) {
-      parrafo.style.display = "block";
-    }
+          deleteTask(item.id);
+        } else if (
+          /* Read more about handling dismissals below */
+          result.dismiss === Swal.DismissReason.cancel
+        ) {
+          swalWithBootstrapButtons.fire(
+            "Cancelled",
+            "Your imaginary file is safe :)",
+            "error"
+          );
+        }
+        let items = ul.getElementsByTagName("li");
+        if (items.length == 0) {
+          parrafo.style.display = "block";
+        }   
+      });
   });
 
   return Eliminar;
@@ -85,17 +121,39 @@ function BtnEliminar() {
 ////////////////////////////////// FUNCION BTN ELIMINAR /////////////////////////////
 
 ////////////////////////////////// FUNCION CHECKBOX /////////////////////////////
-function checkBox() {
+function checkBox(id, isCheck) {
   let btnCheck = document.createElement("input");
   btnCheck.setAttribute("type", "checkbox");
   btnCheck.className = "btnCheckBox";
 
+  btnCheck.checked = isCheck;
+
   btnCheck.addEventListener("click", function () {
+    let check = {
+      check: btnCheck.checked,
+    };
+
+    updateTask(id, check);
+
     if (btnCheck.checked) {
+      Swal.fire({
+        position: "top-end",
+        icon: "success",
+        title: "tarea enviada con exito",
+        showConfirmButton: false,
+        timer: 1000,
+      });
       let cuenta = Number(contador.textContent);
       cuenta = cuenta + 1;
       contador.textContent = cuenta;
     } else if (contador.textContent > 0) {
+      Swal.fire({
+        position: "top-end",
+        icon: "success",
+        title: "tarea quitada con exito",
+        showConfirmButton: false,
+        timer: 1000,
+      });
       let cuenta = Number(contador.textContent);
       cuenta = cuenta - 1;
       contador.textContent = cuenta;
@@ -116,27 +174,29 @@ async function creacion(e) {
       if (texto !== "") {
         input.value = "";
 
-        let task = { task: texto }; // agregar false
+        let task = {
+          task: texto,
+          check: false,
+        };
 
         let resultadoPost = await posTask(task);
 
-        console.log(resultadoPost);
-
         crearTarea(resultadoPost.id, resultadoPost.task);
-
-        parrafo.style.display = "none";
       }
     } else {
-      alert("La tarea que ingresaste esta repetida");
+      Swal.fire("Good job!", "You clicked the button!", "success");
     }
     input.value = "";
   } else {
-    alert("La tarea que ingresaste esta repetida");
+    Swal.fire({
+      icon: "error",
+      title: "OMIGAAA",
+      text: "Hay una tarea repetida chupe mango",
+    });
   }
 }
 
-function crearTarea(id, texto) {
-  //agregar uno mas
+function crearTarea(id, texto, isCheck) {
   let li = document.createElement("li");
   li.className = "li";
   let p = document.createElement("p");
@@ -144,16 +204,25 @@ function crearTarea(id, texto) {
   p.textContent = texto;
 
   ul.appendChild(li);
-  li.appendChild(checkBox()); 
+  li.appendChild(checkBox(id, isCheck));
   li.appendChild(p);
   li.appendChild(BtnEliminar());
+  parrafo.style.display = "none";
 }
 
 async function cargarTareas() {
   let tareas = await getTasks();
+  let contadorCarga = 0;
   tareas.forEach((tarea) => {
-    crearTarea(tarea.id, tarea.task);
+    crearTarea(tarea.id, tarea.task, tarea.check);
   });
+
+  for (let index = 0; index < tareas.length; index++) {
+    if (tareas[index].check == true) {
+      contadorCarga++;
+    }
+  }
+  contador.textContent = contadorCarga;
 }
 
 /////////////////////////////////// FUNCIONES /////////////////////////////////
